Backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-cors==3.0.10
twilio==7.10.0
python-dotenv==1.0.0
gunicorn==20.1.0